feat(console): cloud preview welcome page
