+++ packages/console/src/pages/CloudPreview/cloud-page-size.scss
@@ -1,2 +1,1 @@
 $questionnaire-content-width: 858px;
-$sie-content-width: 1004px;
