--- packages/console/src/pages/CloudPreview/components/ActionBar/actions/WelcomePageAction/index.module.scss
@@ -1,5 +1,0 @@
-@use '@/pages/CloudPreview/cloud-page-size.scss' as size;
-
-.container {
-  max-width: size.$questionnaire-content-width;
-}
