+++ packages/console/src/pages/CloudPreview/components/ActionBar/index.module.scss
@@ -1,12 +1,10 @@
+@use '@/scss/underscore' as _;
+
 .container {
   height: 92px;
   background-color: var(--color-layer-1);
-
-  > div {
-    height: 100%;
-    margin: 0 auto;
-    display: flex;
-    align-items: center;
-    flex-direction: row-reverse;
-  }
+  padding: 0 _.unit(17);
+  display: flex;
+  align-items: center;
+  flex-direction: row-reverse;
 }
