+++ packages/console/src/pages/CloudPreview/components/ActionBar/index.tsx
@@ -1,18 +1,11 @@
-import { useContext } from 'react';
+import type { ReactNode } from 'react';
 
-import { CloudPreviewPageContext } from '../../containers/CloudPreviewPageProvider';
-import { CloudPreviewPage } from '../../types';
-import WelcomePageAction from './actions/WelcomePageAction';
 import * as styles from './index.module.scss';
 
-const ActionBar = () => {
-  const { currentPage } = useContext(CloudPreviewPageContext);
-
-  return (
-    <div className={styles.container}>
-      {currentPage === CloudPreviewPage.Welcome && <WelcomePageAction />}
-    </div>
-  );
+type Props = {
+  children: ReactNode;
 };
+
+const ActionBar = ({ children }: Props) => <div className={styles.container}>{children}</div>;
 
 export default ActionBar;
